+++ lib/nimble_options/validation_error.ex
@@ -2,8 +2,13 @@ defmodule NimbleOptions.ValidationError do
   @moduledoc """
   An error that is returned (or raised) when options are invalid.
 
-  The contents of this exception are not public and you should not rely on
-  its fields.
+  Only these documented fields are considered public. All other fields are
+  considered private and should not be referenced:
+
+    * key: The key that did not successfully validate
+    * keys_path: If the key is nested, this is the path to the key
+    * value: The value that failed to validate. Is set to `nil` if there was no
+      value provided.
 
   Since this is an exception, you can either raise it directly with `raise/1`
   or turn it into a message string with `Exception.message/1`.
@@ -11,7 +16,7 @@ defmodule NimbleOptions.ValidationError do
 
   @type t() :: %__MODULE__{}
 
-  defexception [:message, keys_path: []]
+  defexception [:message, :key, :value, keys_path: []]
 
   @impl true
   def message(%__MODULE__{message: message, keys_path: keys_path}) do